Remove security.User from auditing and split http/grpc/connect
