+++ auditing/async.go
@@ -6,6 +6,7 @@ import (
 	"log/slog"
 	"time"
 
+	"github.com/metal-stack/metal-lib/auditing/api"
 	"github.com/metal-stack/metal-lib/pkg/healthstatus"
 )
 
@@ -24,7 +25,7 @@ type (
 	asyncAuditing struct {
 		log    *slog.Logger
 		config *AsyncConfig
-		a      Auditing
+		a      api.Auditing
 	}
 )
 
@@ -34,7 +35,7 @@ type (
 //
 // Dev note: For a backend wrapped in async, it is strictly required that the index function does not modify internal state
 // as otherwise race conditions will occur!
-func NewAsync(backend Auditing, log *slog.Logger, ac AsyncConfig) (Auditing, error) {
+func NewAsync(backend api.Auditing, log *slog.Logger, ac AsyncConfig) (api.Auditing, error) {
 	wrappedBackendType := fmt.Sprintf("%T", backend)
 
 	a := &asyncAuditing{
@@ -52,7 +53,7 @@ func NewAsync(backend Auditing, log *slog.Logger, ac AsyncConfig) (Auditing, err
 	return a, nil
 }
 
-func (a *asyncAuditing) Index(entry Entry) error {
+func (a *asyncAuditing) Index(entry api.Entry) error {
 	go func() {
 		count := 0
 
@@ -83,7 +84,7 @@ func (a *asyncAuditing) Index(entry Entry) error {
 	return nil
 }
 
-func (a *asyncAuditing) Search(ctx context.Context, filter EntryFilter) ([]Entry, error) {
+func (a *asyncAuditing) Search(ctx context.Context, filter api.EntryFilter) ([]api.Entry, error) {
 	return a.a.Search(ctx, filter)
 }
