+++ auditing/test/auditing_test.go
@@ -1,4 +1,4 @@
-package auditing_test
+package test
 
 import (
 	"context"
@@ -11,11 +11,11 @@ import (
 	"connectrpc.com/connect"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/metal-stack/metal-lib/auditing"
+	"github.com/metal-stack/metal-lib/auditing/api"
+	httpaudit "github.com/metal-stack/metal-lib/auditing/http"
 	"github.com/metal-stack/metal-lib/httperrors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"google.golang.org/grpc/codes"
 )
 
 var (
@@ -26,18 +26,18 @@ var (
 		return x.Unix() == y.Unix()
 	})
 
-	testEntries = func() []auditing.Entry {
-		return []auditing.Entry{
+	testEntries = func() []api.Entry {
+		return []api.Entry{
 			{
-				Component:    "auditing.test",
+				Component:    "test.test",
 				RequestId:    "00000000-0000-0000-0000-000000000000",
-				Type:         auditing.EntryTypeHTTP,
+				Type:         api.EntryTypeHTTP,
 				Timestamp:    now,
 				User:         "admin",
 				Tenant:       "global",
 				Project:      "project",
 				Detail:       "POST",
-				Phase:        auditing.EntryPhaseResponse,
+				Phase:        api.EntryPhaseResponse,
 				Path:         "/v1/test/0",
 				ForwardedFor: "127.0.0.1",
 				RemoteAddr:   "10.0.0.0",
@@ -46,15 +46,15 @@ var (
 				Error:        nil,
 			},
 			{
-				Component:    "auditing.test",
+				Component:    "test.test",
 				RequestId:    "00000000-0000-0000-0000-000000000001",
-				Type:         auditing.EntryTypeHTTP,
+				Type:         api.EntryTypeHTTP,
 				Timestamp:    now.Add(1 * time.Second),
 				User:         "admin",
 				Tenant:       "global",
 				Project:      "project",
 				Detail:       "POST",
-				Phase:        auditing.EntryPhaseResponse,
+				Phase:        api.EntryPhaseResponse,
 				Path:         "/v1/test/1",
 				ForwardedFor: "127.0.0.1",
 				RemoteAddr:   "10.0.0.1",
@@ -63,51 +63,51 @@ var (
 				Error:        nil,
 			},
 			{
-				Component:    "auditing.test",
+				Component:    "test.test",
 				RequestId:    "00000000-0000-0000-0000-000000000002",
-				Type:         auditing.EntryTypeHTTP,
+				Type:         api.EntryTypeHTTP,
 				Timestamp:    now.Add(2 * time.Second),
 				User:         "admin",
 				Tenant:       "global",
 				Project:      "project",
 				Detail:       "POST",
-				Phase:        auditing.EntryPhaseRequest,
+				Phase:        api.EntryPhaseRequest,
 				Path:         "/v1/test/2",
 				ForwardedFor: "127.0.0.1",
 				RemoteAddr:   "10.0.0.2",
 				Body:         "This is the body of 00000000-0000-0000-0000-000000000002",
 				StatusCode:   nil,
-				Error:        auditing.SerializableError(fmt.Errorf("an error")),
+				Error:        httpaudit.SerializableError(fmt.Errorf("an error")),
 			},
 		}
 	}
 
 	tests = func(ctx context.Context) []struct {
 		name string
-		t    func(t *testing.T, a auditing.Auditing)
+		t    func(t *testing.T, a api.Auditing)
 	} {
 		return []struct {
 			name string
-			t    func(t *testing.T, a auditing.Auditing)
+			t    func(t *testing.T, a api.Auditing)
 		}{
 			{
 				name: "no entries, no search results",
-				t: func(t *testing.T, a auditing.Auditing) {
-					entries, err := a.Search(ctx, auditing.EntryFilter{})
+				t: func(t *testing.T, a api.Auditing) {
+					entries, err := a.Search(ctx, api.EntryFilter{})
 					require.NoError(t, err)
 					assert.Empty(t, entries)
 				},
 			},
 			{
 				name: "insert one entry",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{
 						Timestamp: now,
 						Body:      "test",
 					})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						Body: "test",
 					})
 					require.NoError(t, err)
@@ -116,24 +116,24 @@ var (
 			},
 			{
 				name: "insert a couple of entries",
-				t: func(t *testing.T, a auditing.Auditing) {
+				t: func(t *testing.T, a api.Auditing) {
 					es := testEntries()
 					for _, e := range es {
 						err := a.Index(e)
 						require.NoError(t, err)
 					}
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{})
+					entries, err := a.Search(ctx, api.EntryFilter{})
 					require.NoError(t, err)
 					assert.Len(t, entries, len(es))
 
 					sort.Slice(entries, func(i, j int) bool { return entries[i].RequestId < entries[j].RequestId })
 
-					if diff := cmp.Diff(entries, es, cmpopts.IgnoreFields(auditing.Entry{}, "Id", "Error"), timeComparer); diff != "" {
+					if diff := cmp.Diff(entries, es, cmpopts.IgnoreFields(api.Entry{}, "Id", "Error"), timeComparer); diff != "" {
 						t.Errorf("diff (+got -want):\n %s", diff)
 					}
 
-					entries, err = a.Search(ctx, auditing.EntryFilter{
+					entries, err = a.Search(ctx, api.EntryFilter{
 						Body: "This",
 					})
 
@@ -143,15 +143,15 @@ var (
 			},
 			{
 				name: "filter search on component",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{
 						Component: "a",
 						RequestId: "1",
 						Timestamp: now,
 					})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						From:      now.Add(-1 * time.Minute),
 						To:        now.Add(1 * time.Minute),
 						Component: "a",
@@ -159,7 +159,7 @@ var (
 					require.NoError(t, err)
 					require.Len(t, entries, 1)
 
-					if diff := cmp.Diff(entries[0], auditing.Entry{
+					if diff := cmp.Diff(entries[0], api.Entry{
 						Component: "a",
 						RequestId: "1",
 						Timestamp: now,
@@ -170,22 +170,22 @@ var (
 			},
 			{
 				name: "filter search on forwarded for",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{
 						ForwardedFor: "a",
 						RequestId:    "1",
 						Timestamp:    now,
 					})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						ForwardedFor: "a",
 					})
 					require.NoError(t, err)
 					require.Len(t, entries, 1)
 
-					if diff := cmp.Diff(entries[0], auditing.Entry{
-						Component:    "auditing.test",
+					if diff := cmp.Diff(entries[0], api.Entry{
+						Component:    "test.test",
 						ForwardedFor: "a",
 						RequestId:    "1",
 						Timestamp:    now,
@@ -196,22 +196,22 @@ var (
 			},
 			{
 				name: "filter search on path",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{
 						Path:      "/a/b/c/d",
 						RequestId: "1",
 						Timestamp: now,
 					})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						Path: "/b/c", // partial match
 					})
 					require.NoError(t, err)
 					require.Len(t, entries, 1)
 
-					if diff := cmp.Diff(entries[0], auditing.Entry{
-						Component: "auditing.test",
+					if diff := cmp.Diff(entries[0], api.Entry{
+						Component: "test.test",
 						Path:      "/a/b/c/d",
 						RequestId: "1",
 						Timestamp: now,
@@ -222,22 +222,22 @@ var (
 			},
 			{
 				name: "filter search on status code",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{
 						StatusCode: new(400),
 						RequestId:  "1",
 						Timestamp:  now,
 					})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						StatusCode: new(400),
 					})
 					require.NoError(t, err)
 					require.Len(t, entries, 1)
 
-					if diff := cmp.Diff(entries[0], auditing.Entry{
-						Component:  "auditing.test",
+					if diff := cmp.Diff(entries[0], api.Entry{
+						Component:  "test.test",
 						StatusCode: new(400),
 						RequestId:  "1",
 						Timestamp:  now,
@@ -248,28 +248,28 @@ var (
 			},
 			{
 				name: "filter search on status code with 0",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{
 						RequestId: "1",
 						Timestamp: now,
 					})
 					require.NoError(t, err)
 
-					err = a.Index(auditing.Entry{
+					err = a.Index(api.Entry{
 						RequestId:  "2",
 						Timestamp:  now,
 						StatusCode: new(0),
 					})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						StatusCode: new(0),
 					})
 					require.NoError(t, err)
 					require.Len(t, entries, 1)
 
-					if diff := cmp.Diff(entries[0], auditing.Entry{
-						Component:  "auditing.test",
+					if diff := cmp.Diff(entries[0], api.Entry{
+						Component:  "test.test",
 						StatusCode: new(0),
 						RequestId:  "2",
 						Timestamp:  now,
@@ -280,22 +280,22 @@ var (
 			},
 			{
 				name: "filter search on tenant",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{
 						Tenant:    "a",
 						RequestId: "1",
 						Timestamp: now,
 					})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						Tenant: "a",
 					})
 					require.NoError(t, err)
 					require.Len(t, entries, 1)
 
-					if diff := cmp.Diff(entries[0], auditing.Entry{
-						Component: "auditing.test",
+					if diff := cmp.Diff(entries[0], api.Entry{
+						Component: "test.test",
 						Tenant:    "a",
 						RequestId: "1",
 						Timestamp: now,
@@ -306,22 +306,22 @@ var (
 			},
 			{
 				name: "filter search on project",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{
 						Project:   "a",
 						RequestId: "1",
 						Timestamp: now,
 					})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						Project: "a",
 					})
 					require.NoError(t, err)
 					require.Len(t, entries, 1)
 
-					if diff := cmp.Diff(entries[0], auditing.Entry{
-						Component: "auditing.test",
+					if diff := cmp.Diff(entries[0], api.Entry{
+						Component: "test.test",
 						Project:   "a",
 						RequestId: "1",
 						Timestamp: now,
@@ -332,22 +332,22 @@ var (
 			},
 			{
 				name: "filter search on user",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{
 						User:      "a",
 						RequestId: "1",
 						Timestamp: now,
 					})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						User: "a",
 					})
 					require.NoError(t, err)
 					require.Len(t, entries, 1)
 
-					if diff := cmp.Diff(entries[0], auditing.Entry{
-						Component: "auditing.test",
+					if diff := cmp.Diff(entries[0], api.Entry{
+						Component: "test.test",
 						User:      "a",
 						RequestId: "1",
 						Timestamp: now,
@@ -358,14 +358,14 @@ var (
 			},
 			{
 				name: "define a limit higher than amount of results",
-				t: func(t *testing.T, a auditing.Auditing) {
+				t: func(t *testing.T, a api.Auditing) {
 					es := testEntries()
 					for _, e := range es {
 						err := a.Index(e)
 						require.NoError(t, err)
 					}
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						Limit: 1,
 					})
 					require.NoError(t, err)
@@ -378,41 +378,41 @@ var (
 			},
 			{
 				name: "define a limit equal the amount of results",
-				t: func(t *testing.T, a auditing.Auditing) {
+				t: func(t *testing.T, a api.Auditing) {
 					es := testEntries()
 					for _, e := range es {
 						err := a.Index(e)
 						require.NoError(t, err)
 					}
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						Limit: 3,
 					})
 					require.NoError(t, err)
 					require.Len(t, entries, 3)
 
-					if diff := cmp.Diff(entries, es, cmpopts.IgnoreFields(auditing.Entry{}, "Error")); diff != "" {
+					if diff := cmp.Diff(entries, es, cmpopts.IgnoreFields(api.Entry{}, "Error")); diff != "" {
 						t.Errorf("diff (+got -want):\n %s", diff)
 					}
 				},
 			},
 			{
 				name: "define a limit smaller than amount of results",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{
 						RequestId: "1",
 						Timestamp: now,
 					})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						Limit: 5,
 					})
 					require.NoError(t, err)
 					require.Len(t, entries, 1)
 
-					if diff := cmp.Diff(entries[0], auditing.Entry{
-						Component: "auditing.test",
+					if diff := cmp.Diff(entries[0], api.Entry{
+						Component: "test.test",
 						RequestId: "1",
 						Timestamp: now,
 					}); diff != "" {
@@ -422,67 +422,67 @@ var (
 			},
 			{
 				name: "filter search on rqid",
-				t: func(t *testing.T, a auditing.Auditing) {
+				t: func(t *testing.T, a api.Auditing) {
 					es := testEntries()
 					for _, e := range es {
 						err := a.Index(e)
 						require.NoError(t, err)
 					}
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						RequestId: es[0].RequestId,
 					})
 					require.NoError(t, err)
 					require.Len(t, entries, 1)
 
-					if diff := cmp.Diff(entries[0], es[0], cmpopts.IgnoreFields(auditing.Entry{}, "Id", "Error"), timeComparer); diff != "" {
+					if diff := cmp.Diff(entries[0], es[0], cmpopts.IgnoreFields(api.Entry{}, "Id", "Error"), timeComparer); diff != "" {
 						t.Errorf("diff (+got -want):\n %s", diff)
 					}
 				},
 			},
 			{
 				name: "filter search on phase",
-				t: func(t *testing.T, a auditing.Auditing) {
+				t: func(t *testing.T, a api.Auditing) {
 					es := testEntries()
-					var wantEntries []auditing.Entry
+					var wantEntries []api.Entry
 					for _, e := range es {
 						err := a.Index(e)
 						require.NoError(t, err)
 
-						if e.Phase == auditing.EntryPhaseResponse {
+						if e.Phase == api.EntryPhaseResponse {
 							wantEntries = append(wantEntries, e)
 						}
 					}
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
-						Phase: auditing.EntryPhaseResponse,
+					entries, err := a.Search(ctx, api.EntryFilter{
+						Phase: api.EntryPhaseResponse,
 					})
 					require.NoError(t, err)
 					require.Len(t, entries, 2)
 
 					sort.Slice(entries, func(i, j int) bool { return entries[i].RequestId < entries[j].RequestId })
 
-					if diff := cmp.Diff(entries, wantEntries, cmpopts.IgnoreFields(auditing.Entry{}, "Id", "Error"), timeComparer); diff != "" {
+					if diff := cmp.Diff(entries, wantEntries, cmpopts.IgnoreFields(api.Entry{}, "Id", "Error"), timeComparer); diff != "" {
 						t.Errorf("diff (+got -want):\n %s", diff)
 					}
 				},
 			},
 			{
 				name: "filter on body missing one word",
-				t: func(t *testing.T, a auditing.Auditing) {
+				t: func(t *testing.T, a api.Auditing) {
 					es := testEntries()
 					for _, e := range es {
 						err := a.Index(e)
 						require.NoError(t, err)
 					}
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						Body: "This is body",
 					})
 					require.NoError(t, err)
 					assert.Len(t, entries, len(es))
 
-					entries, err = a.Search(ctx, auditing.EntryFilter{
+					entries, err = a.Search(ctx, api.EntryFilter{
 						Body: `"This is body"`,
 					})
 					require.NoError(t, err)
@@ -491,14 +491,14 @@ var (
 			},
 			{
 				name: "filter on body capital ignored",
-				t: func(t *testing.T, a auditing.Auditing) {
+				t: func(t *testing.T, a api.Auditing) {
 					es := testEntries()
 					for _, e := range es {
 						err := a.Index(e)
 						require.NoError(t, err)
 					}
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						Body: "this is the BODY",
 					})
 					require.NoError(t, err)
@@ -507,14 +507,14 @@ var (
 			},
 			{
 				name: "filter on body",
-				t: func(t *testing.T, a auditing.Auditing) {
+				t: func(t *testing.T, a api.Auditing) {
 					es := testEntries()
 					for _, e := range es {
 						err := a.Index(e)
 						require.NoError(t, err)
 					}
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						Body: es[0].Body.(string),
 					})
 					require.NoError(t, err)
@@ -527,38 +527,38 @@ var (
 			},
 			{
 				name: "filter on body partial words",
-				t: func(t *testing.T, a auditing.Auditing) {
+				t: func(t *testing.T, a api.Auditing) {
 					es := testEntries()
 					for _, e := range es {
 						err := a.Index(e)
 						require.NoError(t, err)
 					}
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						Body: "002",
 					})
 					require.NoError(t, err)
 					require.Len(t, entries, 1)
 
-					if diff := cmp.Diff(entries[0], es[2], cmpopts.IgnoreFields(auditing.Entry{}, "Error")); diff != "" {
+					if diff := cmp.Diff(entries[0], es[2], cmpopts.IgnoreFields(api.Entry{}, "Error")); diff != "" {
 						t.Errorf("diff (+got -want):\n %s", diff)
 					}
 				},
 			},
 			{
 				name: "filter on every filter field",
-				t: func(t *testing.T, a auditing.Auditing) {
+				t: func(t *testing.T, a api.Auditing) {
 					es := testEntries()
 					for _, e := range es {
 						err := a.Index(e)
 						require.NoError(t, err)
 					}
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						Limit:        1,
 						From:         now.Add(-1 * time.Minute),
 						To:           now.Add(1 * time.Minute),
-						Component:    "auditing.test",
+						Component:    "test.test",
 						RequestId:    "00000000-0000-0000-0000-000000000000",
 						Type:         "http",
 						User:         "admin",
@@ -583,28 +583,28 @@ var (
 			},
 			{
 				name: "filter on nothing",
-				t: func(t *testing.T, a auditing.Auditing) {
+				t: func(t *testing.T, a api.Auditing) {
 					es := testEntries()
 					for _, e := range es {
 						err := a.Index(e)
 						require.NoError(t, err)
 					}
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{})
+					entries, err := a.Search(ctx, api.EntryFilter{})
 					require.NoError(t, err)
 					require.Len(t, entries, len(testEntries()))
 				},
 			},
 			{
 				name: "filter on query does not affect other filters",
-				t: func(t *testing.T, a auditing.Auditing) {
+				t: func(t *testing.T, a api.Auditing) {
 					es := testEntries()
 					for _, e := range es {
 						err := a.Index(e)
 						require.NoError(t, err)
 					}
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{
+					entries, err := a.Search(ctx, api.EntryFilter{
 						From:      now.Add(-1 * time.Minute),
 						To:        now.Add(1 * time.Minute),
 						RequestId: "00000000-0000-0000-0000-000000000000",
@@ -622,46 +622,46 @@ var (
 			},
 			{
 				name: "fields are defaulted during indexing",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{})
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{})
+					entries, err := a.Search(ctx, api.EntryFilter{})
 					require.NoError(t, err)
 					assert.Len(t, entries, 1)
-					assert.Equal(t, "auditing.test", entries[0].Component)
+					assert.Equal(t, "test.test", entries[0].Component)
 					assert.WithinDuration(t, time.Now(), entries[0].Timestamp, 1*time.Second)
 				},
 			},
 			{
 				name: "index an http error",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{
-						Error: auditing.SerializableError(httperrors.NewHTTPError(http.StatusConflict, fmt.Errorf("already exists"))),
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{
+						Error: httpaudit.SerializableError(httperrors.NewHTTPError(http.StatusConflict, fmt.Errorf("already exists"))),
 					})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{})
+					entries, err := a.Search(ctx, api.EntryFilter{})
 					require.NoError(t, err)
 					assert.Len(t, entries, 1)
-					assert.Equal(t, "auditing.test", entries[0].Component)
+					assert.Equal(t, "test.test", entries[0].Component)
 					assert.Equal(t, map[string]any{"statuscode": float64(409), "message": "already exists"}, entries[0].Error)
 					assert.WithinDuration(t, time.Now(), entries[0].Timestamp, 1*time.Second)
 				},
 			},
 			{
 				name: "index a connect error",
-				t: func(t *testing.T, a auditing.Auditing) {
-					err := a.Index(auditing.Entry{
-						Error: auditing.SerializableError(connect.NewError(connect.CodeAlreadyExists, fmt.Errorf("already exists"))),
+				t: func(t *testing.T, a api.Auditing) {
+					err := a.Index(api.Entry{
+						Error: httpaudit.SerializableError(connect.NewError(connect.CodeAlreadyExists, fmt.Errorf("already exists"))),
 					})
 					require.NoError(t, err)
 
-					entries, err := a.Search(ctx, auditing.EntryFilter{})
+					entries, err := a.Search(ctx, api.EntryFilter{})
 					require.NoError(t, err)
 					assert.Len(t, entries, 1)
-					assert.Equal(t, "auditing.test", entries[0].Component)
-					assert.Equal(t, map[string]any{"code": float64(codes.AlreadyExists), "error": "already_exists: already exists", "message": "already_exists"}, entries[0].Error)
+					assert.Equal(t, "test.test", entries[0].Component)
+					assert.Equal(t, map[string]any{"error": "already_exists: already exists"}, entries[0].Error)
 					assert.WithinDuration(t, time.Now(), entries[0].Timestamp, 1*time.Second)
 				},
 			},
