+++ auditing/test/memory_test.go
@@ -1,4 +1,4 @@
-package auditing_test
+package test
 
 import (
 	"context"
@@ -7,7 +7,8 @@ import (
 	"slices"
 	"testing"
 
-	"github.com/metal-stack/metal-lib/auditing"
+	"github.com/metal-stack/metal-lib/auditing/api"
+	"github.com/metal-stack/metal-lib/auditing/memory"
 	"github.com/metal-stack/metal-lib/pkg/healthstatus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -25,18 +26,18 @@ func TestAuditing_Memory(t *testing.T) {
 				t.Skipf("skipping because memory backend does not support this")
 			}
 
-			auditing, err := auditing.NewMemory(auditing.Config{
+			auditing, err := memory.NewMemory(api.Config{
 				Log: slog.Default(),
-			}, auditing.MemoryConfig{})
+			}, memory.MemoryConfig{})
 			require.NoError(t, err)
 
 			tt.t(t, auditing)
 		})
 	}
 
-	auditing, err := auditing.NewMemory(auditing.Config{
+	auditing, err := memory.NewMemory(api.Config{
 		Log: slog.Default(),
-	}, auditing.MemoryConfig{})
+	}, memory.MemoryConfig{})
 	require.NoError(t, err)
 
 	healthResult, err := auditing.Check(t.Context())
